requirements.txt
discord
python-dotenv
mpi4py
